test(compose): cover queue slot picker selection behavior
Rewrite queue-preview test as a .tsx component test that renders
QueuePreview and simulates picking a slot via the Base UI Select,
asserting onSelectSlot receives the chosen slot's ISO timestamp,
instead of grepping source files for select markup.
